Moves extension method to correct location
Moves the extension method for replacing the model in GenerativeAIChatClient to a more appropriate location, and renames the class to reflect its purpose as an extension for the GenerativeAIChatClient.
